add LoggerWriter class for stdout/stderr redirection to logging system
